rename logcol package to logd
