+++ cmd/logp/main.go
@@ -6,8 +6,8 @@ import (
 	"context"
 
 	"github.com/deref/exo/cmdutil"
-	"github.com/deref/exo/logcol/api"
-	"github.com/deref/exo/logcol/server"
+	"github.com/deref/exo/logd/api"
+	"github.com/deref/exo/logd/server"
 	"github.com/deref/pier"
 )
