+++ logd/api/collector.go
