+++ logd/api/consts.go
