+++ logd/server/badger.go
