+++ logd/server/id.go
