+++ logd/server/state.go
