+++ logd/server/workers.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/deref/exo/chrono"
-	"github.com/deref/exo/logcol/api"
+	"github.com/deref/exo/logd/api"
 )
 
 type worker struct {
